deployed contract to sepolia ethereum test net
- changed the node URL of the network
- changed the path to the tintolen.json

+++ backend/E_commerce_BC/api/models.py
@@ -62,7 +62,7 @@ class ProductItem(models.Model):
     product_image = models.CharField(null=True , max_length=1000)
     IPFS_hash = models.URLField(null= True , max_length=1000)
     img_url = models.URLField(null=True , max_length=1000)
-    prize = models.IntegerField()
+    prize = models.FloatField()
     added_on = models.DateTimeField(null=True)
 
     def save(self, *args, **kwargs):
